Rename to json_parser to correct json_selector_parser
