+++ ndjson-common/src/lib.rs
@@ -17,5 +17,5 @@
 pub mod common;
 pub mod error;
 pub mod from;
-pub mod json_parser;
+pub mod json_selector_parser;
 pub mod ndjson;
